Move version to its own package. (#3460)
