+++ checkpoint.go
@@ -8,6 +8,7 @@ import (
 	"github.com/hashicorp/go-checkpoint"
 	"github.com/mitchellh/packer/command"
 	"github.com/mitchellh/packer/packer"
+	packerVersion "github.com/mitchellh/packer/version"
 )
 
 func init() {
@@ -33,9 +34,9 @@ func runCheckpoint(c *config) {
 		return
 	}
 
-	version := Version
-	if VersionPrerelease != "" {
-		version += fmt.Sprintf("-%s", VersionPrerelease)
+	version := packerVersion.Version
+	if packerVersion.VersionPrerelease != "" {
+		version += fmt.Sprintf("-%s", packerVersion.VersionPrerelease)
 	}
 
 	signaturePath := filepath.Join(configDir, "checkpoint_signature")
