+++ commands.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/mitchellh/cli"
 	"github.com/mitchellh/packer/command"
+	"github.com/mitchellh/packer/version"
 )
 
 // Commands is the mapping of all the available Terraform commands.
@@ -53,9 +54,9 @@ func init() {
 		"version": func() (cli.Command, error) {
 			return &command.VersionCommand{
 				Meta:              *CommandMeta,
-				Revision:          GitCommit,
-				Version:           Version,
-				VersionPrerelease: VersionPrerelease,
+				Revision:          version.GitCommit,
+				Version:           version.Version,
+				VersionPrerelease: version.VersionPrerelease,
 				CheckFunc:         commandVersionCheck,
 			}, nil
 		},
