+++ version/version.go
@@ -1,4 +1,9 @@
-package main
+package version
+
+import (
+	"bytes"
+	"fmt"
+)
 
 // The git commit that was compiled. This will be filled in by the compiler.
 var GitCommit string
@@ -10,3 +15,17 @@ const Version = "0.10.1"
 // then it means that it is a final release. Otherwise, this is a pre-release
 // such as "dev" (in development), "beta", "rc1", etc.
 const VersionPrerelease = "dev"
+
+func FormattedVersion() string {
+	var versionString bytes.Buffer
+	fmt.Fprintf(&versionString, "%s", Version)
+	if VersionPrerelease != "" {
+		fmt.Fprintf(&versionString, ".%s", VersionPrerelease)
+
+		if GitCommit != "" {
+			fmt.Fprintf(&versionString, " (%s)", GitCommit)
+		}
+	}
+
+	return versionString.String()
+}
